v2/resumePDF/requirements.txt
# pip install -r requirements.txt
requests
beautifulsoup4
selenium
PyPDF2
reportlab
pdfplumber
spacy
PyQt5
mysql-connector-python
lxml